+++ Sources/TuistEnvKit/Commands/UpdateCommand.swift
@@ -9,14 +9,7 @@ struct UpdateCommand: ParsableCommand {
                              abstract: "Installs the latest version if it's not already installed")
     }
 
-    /// Force argument (-f). When passed, it re-installs the latest version compiling it from the source.
-    @Flag(
-        name: .shortAndLong,
-        help: "Re-installs the latest version compiling it from the source"
-    )
-    var force: Bool = false
-
     func run() throws {
-        try UpdateService().run(force: force)
+        try UpdateService().run()
     }
 }