+++ Sources/TuistGenerator/Linter/GraphLinter.swift
@@ -342,6 +342,11 @@ public class GraphLinter: GraphLinting {
             LintableTarget(platform: .iOS, product: .dynamicLibrary),
             LintableTarget(platform: .iOS, product: .framework),
         ],
+        LintableTarget(platform: .iOS, product: .appClip): [
+            LintableTarget(platform: .iOS, product: .staticLibrary),
+            LintableTarget(platform: .iOS, product: .dynamicLibrary),
+            LintableTarget(platform: .iOS, product: .framework),
+        ],
         //        LintableTarget(platform: .iOS, product: .messagesApplication): [
 //            LintableTarget(platform: .iOS, product: .messagesExtension),
 //            LintableTarget(platform: .iOS, product: .staticLibrary),