+++ Sources/TuistLoader/Models+ManifestMappers/RunAction+ManifestMapper.swift
@@ -20,11 +20,16 @@ extension TuistGraph.RunAction {
             executableResolved = TargetReference(projectPath: try generatorPaths.resolveSchemeActionProjectPath(executable.projectPath),
                                                  name: executable.targetName)
         }
+
+        let options = try TuistGraph.RunActionOptions.from(manifest: manifest.options, generatorPaths: generatorPaths)
+
         let diagnosticsOptions = Set(manifest.diagnosticsOptions.map { TuistGraph.SchemeDiagnosticsOption.from(manifest: $0) })
+
         return TuistGraph.RunAction(configurationName: configurationName,
                                     executable: executableResolved,
                                     filePath: nil,
                                     arguments: arguments,
+                                    options: options,
                                     diagnosticsOptions: diagnosticsOptions)
     }
 }